+++ src/SubPopupMenu.js
@@ -8,10 +8,10 @@ const SubPopupMenu = React.createClass({
     onSelect: React.PropTypes.func,
     onClick: React.PropTypes.func,
     onDeselect: React.PropTypes.func,
-    onExpandedChange: React.PropTypes.func,
+    onOpenedChange: React.PropTypes.func,
     onDestroy: React.PropTypes.func,
-    expandedKeys: React.PropTypes.arrayOf(React.PropTypes.string),
-    openSubMenuOnMouseEnter: React.PropTypes.bool,
+    openedKeys: React.PropTypes.arrayOf(React.PropTypes.string),
+    closeSubMenuOnMouseLeave: React.PropTypes.bool,
   },
 
   mixins: [MenuMixin],
@@ -28,45 +28,27 @@ const SubPopupMenu = React.createClass({
     this.props.onClick(e);
   },
 
-  onExpandedChange(e) {
-    this.props.onExpandedChange(e);
-  },
-
-  onItemHover({hover, key, item, trigger}) {
-    if (!hover && trigger) {
-      if (!this.props.openSubMenuOnMouseEnter) {
-        return;
-      }
-    }
-    this.setState({
-      activeKey: hover ? key : null,
-    });
-
-    if (hover && !item.isSubMenu) {
-      const subMenu = this.lastExpandedSubMenu();
-      if (subMenu && key !== subMenu.props.eventKey) {
-        this.onExpandedChange({
-          key: subMenu.props.eventKey,
-          expanded: false,
-          item: subMenu,
-          trigger: 'mouseleave',
-        });
-      }
-    }
+  onOpenedChange(e) {
+    this.props.onOpenedChange(e);
   },
 
   onDestroy(key) {
     this.props.onDestroy(key);
   },
 
+  onItemHover(e) {
+    this.onCommonItemHover(e);
+  },
+
   renderMenuItem(c, i) {
     const key = getKeyFromChildrenIndex(c, i);
     const props = this.props;
     const extraProps = {
-      expandedKeys: props.expandedKeys,
+      openedKeys: props.openedKeys,
       selectedKeys: props.selectedKeys,
-      expanded: props.expandedKeys.indexOf(key) !== -1,
+      opened: props.openedKeys.indexOf(key) !== -1,
       selected: props.selectedKeys.indexOf(key) !== -1,
+      openSubMenuOnMouseEnter: true,
     };
     return this.renderCommonMenuItem(c, i, extraProps);
   },
@@ -75,16 +57,6 @@ const SubPopupMenu = React.createClass({
     const props = assign({}, this.props);
     props.className += ` ${props.prefixCls}-sub`;
     return this.renderRoot(props);
-  },
-
-  lastExpandedSubMenu() {
-    let lastOpen = [];
-    if (this.props.expandedKeys.length) {
-      lastOpen = this.instanceArray.filter((c)=> {
-        return this.props.expandedKeys.indexOf(c.props.eventKey) !== -1;
-      });
-    }
-    return lastOpen[0];
   },
 });